+++ tensorflow/lite/micro/kernels/add.cc
@@ -77,15 +77,14 @@ TfLiteStatus CalculateOpData(TfLiteContext* context, TfLiteAddParams* params,
     data->output_offset = output->params.zero_point;
     data->left_shift = 20;
     const double twice_max_input_scale =
-        2 * static_cast<double>(
-                std::max(input1->params.scale, input2->params.scale));
+        2 * std::max(input1->params.scale, input2->params.scale);
     const double real_input1_multiplier =
-        static_cast<double>(input1->params.scale) / twice_max_input_scale;
+        input1->params.scale / twice_max_input_scale;
     const double real_input2_multiplier =
-        static_cast<double>(input2->params.scale) / twice_max_input_scale;
+        input2->params.scale / twice_max_input_scale;
     const double real_output_multiplier =
         twice_max_input_scale /
-        ((1 << data->left_shift) * static_cast<double>(output->params.scale));
+        ((1 << data->left_shift) * output->params.scale);
 
     QuantizeMultiplierSmallerThanOneExp(
         real_input1_multiplier, &data->input1_multiplier, &data->input1_shift);
